:hammer: Devide cli entry and logic
